Broke up Project Page into components
