+++ app/src/components/projects/ProjectCarouselItem.css
@@ -1,0 +1,24 @@
+.carousel-item-content {
+    position: relative;
+    width: 100%;
+    height: 400px; /* Adjust height as needed */
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    color: white;
+    text-shadow: 2px 2px 4px rgba(0, 0, 0, 0.7);
+    background-color: #2b3035;
+    border: 1px solid #f8f9fa;
+  }
+  
+  [data-bs-theme="light"] .carousel-item-content {
+    background-color: #f8f9fa;
+    color: #2b3035;
+    border: 1px solid #2b3035;
+  }
+  
+  .carousel-item-content h3,
+  .carousel-item-content p {
+    margin: 0;
+  }
+
